requirements.txt
oci
python-dotenv
requests
croniter 